# cherrypick (#172): internal/lz4block: arm64 decoder improvements