Pulls in 7c0f44d5dc8a9d0db8a06ce7b2a3a6b5484a5cf4 and 24303cffaad515ae9d6613a91a0f725003b1847d